# frontend/public/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fish-Tech バックエンド API 仕様です。",
        "title": "Fish-Tech API",
        "contact": {},
        "version": "1.1"
    },
    "basePath": "/api",
    "paths": {
...
        "/fishes": {
            "get": {
                "description": "公開中の魚一覧を返します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "魚一覧を取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PublicFishListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/pairs": {
            "get": {
                "description": "公開中の魚相性一覧を返します。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "魚相性一覧を取得",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PublicPairListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.PublicFishListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.PublicFishResponse"
                    }
                }
            }
        },
        "handler.PublicFishResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "linkUrl": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handler.PublicPairListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.PublicPairResponse"
                    }
                }
            }
        },
        "handler.PublicPairResponse": {
            "type": "object",
            "properties": {
                "fishIdA": {
                    "type": "string"
                },
                "fishIdB": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "memo": {
                    "type": "string"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
...
    }
}